Move archiving into useRule hook alongside other API methods
