--- apps/rule-manager/client/src/ts/components/api/archiveRule.ts
@@ -1,7 +1,0 @@
-import { responseHandler } from "./parseResponse";
-
-export const archiveRule = async (ruleId: number) => (
-    fetch(`${location}rules/${ruleId}/archive`, {
-        method: 'POST',
-    }).then(responseHandler)
-);
